style: rename files to kebab-case and adjust spacing for readability (web)
