+++ apps/web/components/landing-page.tsx
@@ -1,9 +1,10 @@
 import Image from "next/image";
 import Link from "next/link";
 import type { FC } from "react";
+import { Button } from "@repo/ui/button";
 
-const LandingPage: FC = () => {
-  return (
+const LandingPage: FC = () => (
+
     <div className="min-h-screen bg-gradient-to-br from-slate-50 to-blue-50">
 
       {/* Header */}
@@ -18,9 +19,10 @@ const LandingPage: FC = () => {
           />
           <span className="text-xl font-bold text-gray-900">TwitterClone</span>
         </div>
-        <button className="text-[#00AAEC] hover:text-[#1DA1F2] font-medium transition-colors">
+
+        <Button className="text-[#00AAEC] hover:text-[#1DA1F2] font-medium transition-colors" appName="web">
           I have an account
-        </button>
+        </Button>
       </header>
 
       {/* Hero Section */}
@@ -34,6 +36,7 @@ const LandingPage: FC = () => {
                 Connect with the world in{" "}
                 <span className="text-[#00AAEC]">real time</span>
               </h1>
+
               <p className="text-xl text-gray-600 max-w-lg">
                 Join millions of people sharing thoughts, ideas, and moments that matter to them.
               </p>
@@ -51,6 +54,7 @@ const LandingPage: FC = () => {
           {/* Right Illustration */}
           <div className="relative lg:order-last order-first">
             <div className="relative bg-gradient-to-br from-[#00AAEC]/10 to-[#1DA1F2]/10 rounded-3xl p-8 backdrop-blur-sm">
+
               {/* Mock Tweet Cards */}
               <div className="space-y-4">
                 <div className="bg-white rounded-2xl p-6 shadow-lg hover:shadow-xl transition-shadow">
@@ -58,14 +62,17 @@ const LandingPage: FC = () => {
                     <div className="w-12 h-12 bg-gradient-to-r from-[#00AAEC] to-[#1DA1F2] rounded-full flex items-center justify-center">
                       <span className="text-white font-bold text-lg">A</span>
                     </div>
+
                     <div>
                       <div className="font-semibold text-gray-900">Alex Chen</div>
                       <div className="text-gray-500 text-sm">@alexc</div>
                     </div>
                   </div>
+
                   <p className="text-gray-700">
                     Just shipped a new feature! The community response has been incredible 🚀
                   </p>
+
                   <div className="flex items-center gap-6 mt-4 text-gray-400 text-sm">
                     <span className="hover:text-[#00AAEC] cursor-pointer">💬 12</span>
                     <span className="hover:text-green-500 cursor-pointer">🔄 8</span>
@@ -83,9 +90,11 @@ const LandingPage: FC = () => {
                       <div className="text-gray-500 text-sm">@maria_dev</div>
                     </div>
                   </div>
+
                   <p className="text-gray-700">
                     Love how this platform brings developers together! 💻✨
                   </p>
+
                   <div className="flex items-center gap-6 mt-4 text-gray-400 text-sm">
                     <span className="hover:text-[#00AAEC] cursor-pointer">💬 5</span>
                     <span className="hover:text-green-500 cursor-pointer">🔄 15</span>
@@ -108,6 +117,5 @@ const LandingPage: FC = () => {
       </footer>
     </div>
   );
-};
 
 export default LandingPage;
